Prototype: autocomplete search box and domain status table
The domain view now includes both the vanilla autocomplete directive
for domain searches and a table that shows more of the domain search
result data set.

The intention is to demonstrate the vanilla directive and also start
to work with the actual data that should be combined to display the
search results.