Fixed template upload and handle deleted site folders
There was a logic error due to poor indentation that caused
templates to be uploaded even if the website model was not persisted
(e.g. due to quota limits). Not expected to have effected any users.

Additionally, while investigating the above it became clear we were
not handling the Dropbox metadata response object for folders that once
existed but are now deleted (e.g. user deletes website folder and
re-creates it).